Python-grpc/requirements.txt
flask
pymongo
redis
grpcio
grpcio-tools
